requirements.txt
Flask==2.0.1
flask-cors==3.0.10
pyserial==3.5
